Correctly cast integers

+++ testdata/test1.go
@@ -12,8 +12,8 @@ func TestY() (int8, int16) {
 }
 */
 
-func Test(a, b int32, c int64) int32 {
-	var aa, bc int32 = b*int32(c), a
+func Test(a, b int64, c int32) int64 {
+	var aa, bc int64 = b*int64(c), a
 	aa, bc = bc, aa
 	return (aa + bc / a) % a
 }
